requirements.txt
openai>=1.40.0
requests>=2.31.0
tqdm>=4.66.0
tenacity>=8.2.3
genanki>=0.13.1
google-cloud-texttospeech>=2.16.3
google-generativeai>=0.7.2
google-cloud-translate>=3.15.4
